Standardization of files names
